python-pub/requirements.txt
asyncio==3.4.3
nats_py==2.6.0